+++ pkg/application/mock_componentInstallationUseCase_test.go
@@ -5,7 +5,9 @@ package application
 import (
 	context "context"
 
+	domain "github.com/cloudogu/k8s-blueprint-operator/pkg/domain"
 	ecosystem "github.com/cloudogu/k8s-blueprint-operator/pkg/domain/ecosystem"
+
 	mock "github.com/stretchr/testify/mock"
 )
 
@@ -165,6 +167,50 @@ func (_c *mockComponentInstallationUseCase_WaitForHealthyComponents_Call) Return
 }
 
 func (_c *mockComponentInstallationUseCase_WaitForHealthyComponents_Call) RunAndReturn(run func(context.Context) (ecosystem.ComponentHealthResult, error)) *mockComponentInstallationUseCase_WaitForHealthyComponents_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
+// applyComponentState provides a mock function with given fields: _a0, _a1, _a2
+func (_m *mockComponentInstallationUseCase) applyComponentState(_a0 context.Context, _a1 domain.ComponentDiff, _a2 *ecosystem.ComponentInstallation) error {
+	ret := _m.Called(_a0, _a1, _a2)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(context.Context, domain.ComponentDiff, *ecosystem.ComponentInstallation) error); ok {
+		r0 = rf(_a0, _a1, _a2)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// mockComponentInstallationUseCase_applyComponentState_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'applyComponentState'
+type mockComponentInstallationUseCase_applyComponentState_Call struct {
+	*mock.Call
+}
+
+// applyComponentState is a helper method to define mock.On call
+//   - _a0 context.Context
+//   - _a1 domain.ComponentDiff
+//   - _a2 *ecosystem.ComponentInstallation
+func (_e *mockComponentInstallationUseCase_Expecter) applyComponentState(_a0 interface{}, _a1 interface{}, _a2 interface{}) *mockComponentInstallationUseCase_applyComponentState_Call {
+	return &mockComponentInstallationUseCase_applyComponentState_Call{Call: _e.mock.On("applyComponentState", _a0, _a1, _a2)}
+}
+
+func (_c *mockComponentInstallationUseCase_applyComponentState_Call) Run(run func(_a0 context.Context, _a1 domain.ComponentDiff, _a2 *ecosystem.ComponentInstallation)) *mockComponentInstallationUseCase_applyComponentState_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(context.Context), args[1].(domain.ComponentDiff), args[2].(*ecosystem.ComponentInstallation))
+	})
+	return _c
+}
+
+func (_c *mockComponentInstallationUseCase_applyComponentState_Call) Return(_a0 error) *mockComponentInstallationUseCase_applyComponentState_Call {
+	_c.Call.Return(_a0)
+	return _c
+}
+
+func (_c *mockComponentInstallationUseCase_applyComponentState_Call) RunAndReturn(run func(context.Context, domain.ComponentDiff, *ecosystem.ComponentInstallation) error) *mockComponentInstallationUseCase_applyComponentState_Call {
 	_c.Call.Return(run)
 	return _c
 }