+++ pkg/domain/stateDiffComponent.go
@@ -13,8 +13,8 @@ import (
 type ComponentDiffs []ComponentDiff
 
 // Statistics aggregates various figures about the required actions of the ComponentDiffs.
-func (cd ComponentDiffs) Statistics() (toInstall int, toUpgrade int, toUninstall int, other int) {
-	for _, componentDiff := range cd {
+func (diffs ComponentDiffs) Statistics() (toInstall int, toUpgrade int, toUninstall int, other int) {
+	for _, componentDiff := range diffs {
 		switch componentDiff.NeededAction {
 		case ActionInstall:
 			toInstall += 1
@@ -27,6 +27,16 @@ func (cd ComponentDiffs) Statistics() (toInstall int, toUpgrade int, toUninstall
 		}
 	}
 	return
+}
+
+// GetComponentDiffByName returns the diff for the given component name or an empty struct if it was not found.
+func (diffs ComponentDiffs) GetComponentDiffByName(name common.SimpleComponentName) ComponentDiff {
+	for _, diff := range diffs {
+		if diff.Name == name {
+			return diff
+		}
+	}
+	return ComponentDiff{}
 }
 
 // ComponentDiff represents the Diff for a single expected Component to the current ecosystem.ComponentInstallation.
@@ -53,8 +63,12 @@ type ComponentDiffState struct {
 	InstallationState TargetState
 }
 
+func (diff ComponentDiff) HasChanges() bool {
+	return diff.NeededAction != ActionNone
+}
+
 // String returns a string representation of the ComponentDiff.
-func (diff *ComponentDiff) String() string {
+func (diff ComponentDiff) String() string {
 	return fmt.Sprintf(
 		"{Name: %q, Actual: %s, Expected: %s, NeededAction: %q}",
 		diff.Name,
@@ -65,7 +79,7 @@ func (diff *ComponentDiff) String() string {
 }
 
 // String returns a string representation of the ComponentDiffState.
-func (diff *ComponentDiffState) String() string {
+func (diff ComponentDiffState) String() string {
 	return fmt.Sprintf(
 		"{Namespace: %q, Version: %q, InstallationState: %q}",
 		diff.Namespace,
@@ -74,7 +88,7 @@ func (diff *ComponentDiffState) String() string {
 	)
 }
 
-func (diff *ComponentDiffState) getSafeVersionString() string {
+func (diff ComponentDiffState) getSafeVersionString() string {
 	if diff.Version != nil {
 		return diff.Version.String()
 	} else {